Problem Resolved(?)
There was a problem that was preventing the program from running properly because of a copy constructor in the Weapons class. This has been fixed, possibly.

These extra files are just the SpriteLib.h and SpriteLib.cpp. Most of them won't be changed, but this just makes copying things easier.